fix sam & add clip_point_only

+++ core/services/mask_service.py
@@ -1,6 +1,7 @@
 #mask_service.py
 from typing import List
 from initialize_model import decoder
+import numpy as np
 
 # 一个点
 # _, logits = decoder.point(img_embeddings, img_file, point_coords=[[324, 282]], point_labels=[1])
@@ -24,16 +25,16 @@ def generate_mask(operation, current_state, img_embeddings, img_file, logits):
         #101 111 011
         if (current_state["foreground"] or current_state["background"]) and current_state["boxes"]:
             current_lables = [1] * len(current_state["foreground"]) + [0] * len(current_state["background"])
-            masks, logits = decoder.hybrid(img_embeddings, img_file, point_coords=current_state["foreground"]+current_state["background"], point_labels=current_lables, boxes=current_state["boxes"], logits=logits)
+            masks, logits = decoder.hybrid(img_embeddings.astype(np.float32), img_file, point_coords=current_state["foreground"]+current_state["background"], point_labels=current_lables, boxes=current_state["boxes"], logits=logits)
         #100 110 010    
         elif (current_state["foreground"] or current_state["background"]) and not current_state["boxes"]:
             current_lables = [1] * len(current_state["foreground"]) + [0] * len(current_state["background"])
-            masks, logits = decoder.point(img_embeddings, img_file, point_coords=current_state["foreground"]+current_state["background"], point_labels=current_lables, logits=logits)
+            masks, logits = decoder.point(img_embeddings.astype(np.float32), img_file, point_coords=current_state["foreground"]+current_state["background"], point_labels=current_lables, logits=logits)
         #001
         else:
             if not current_state["boxes"]:
                 raise ValueError("Boxes is not None")
-            masks, logits = decoder.bBox(img_embeddings, img_file, boxes=current_state["boxes"])
+            masks, logits = decoder.bBox(img_embeddings.astype(np.float32), img_file, boxes=current_state["boxes"])
         masks_2d = masks.astype(int).squeeze().tolist()
         
     return masks_2d, logits
